♻️ localization update:
Unfortunately the <if> element does not render outside of a field in a form and therefore we must use different files to enforce language localization until we deploy the app on a custom server.
The folder structure has been changed to host both english and french version of the application.

For the same reason, we cannot use application variables to render different prompts and therefore we must use multiple copies of the subcategory section for now.